// src/types.ts
export interface Physics {
  packetId: number;
  gas: number;
  brake: number;
  fuel: number;
  gear: number;
  rpms: number;
  steerAngle: number;
  kmh: number;
  velocity: number[];
  accG: number[];
  wheelSlip: number[];
  wheelLoad: number[];
  wheelPressure: number[];
  wheelAngularSpeed: number[];
  tyreWear: number[];
  tyreDirtyLevel: number[];
  tyreCoreTemp: number[];
  camberRAD: number[];
  suspensionTravel: number[];
  drs: number;
  tc: number;
  heading: number;
  pitch: number;
  roll: number;
  cgHeight: number;
  carDamage: number[];
  numberOfTyresOut: number;
  isPitLimiterOn: boolean;
  abs: number;
  kersChange: number;
  kersInput: number;
  isAutoShifterOn: boolean;
  rideHeight: number[];
  turboBoost: number;
  ballast: number;
  airDensity: number;
  airTemp: number;
  roadTemp: number;
  localAngularVel: number[];
  finalFF: number;
  performanceMeter: number;
  engineBrake: number;
  ersRecoveryLevel: number;
  ersPowerLevel: number;
  ersHeatCharging: number;
  isERSCharging: boolean;
  kersCurrentKJ: number;
  isDRSAvailable: boolean;
  isDRSEnabled: boolean;
  brakeTemp: number[];
  clutch: number;
  tyreTempI: number[];
  tyreTempM: number[];
  tyreTempO: number[];
  isAIControlled: boolean;
  tyreContactPoint: Array<number[]>;
  tyreContactNormal: Array<number[]>;
  tyreContactHeading: Array<number[]>;
  brakeBias: number;
  localVelocity: number[];
  p2pActivation: number;
  p2pStatus: number;
  currentMaxRpm: number;
  mz: number[];
  fx: number[];
  fy: number[];
  slipRatio: number[];
  slipAngle: number[];
  tcinAction: number;
  absInAction: number;
  suspensionDamage: number[];
  tyreTemp: number[];
  waterTemp: number;
  brakePressure: number[];
  frontBrakeCompound: number;
  rearBrakeCompound: number;
  padLife: number[];
  discLife: number[];
  isIgnitionOn: boolean;
  isEngineStarterOn: boolean;
  isEngineRunning: boolean;
  kerbVibration: number;
  slipVibrations: number;
  gVibrations: number;
  absVibrations: number;
}

type RainIntensity =
  | 'no-rain'
  | 'drizzle'
  | 'light-rain'
  | 'medium-rain'
  | 'heavy-rain'
  | 'thunderstorm';

export interface Graphics {
  packetId: number;
  status: 'off' | 'replay' | 'live' | 'pause';
  session:
    | 'unknown'
    | 'practice'
    | 'qualifying'
    | 'race'
    | 'hot-lap'
    | 'time-attack'
    | 'drift'
    | 'drag'
    | 'hot-stint'
    | 'hot-lap-super-pole';
  currentTime: string;
  lastTime: string;
  bestTime: string;
  split: string;
  completedLaps: number;
  position: number;
  iCurrentTime: number;
  iLastTime: number;
  iBestTime: number;
  sessionTimeLeft: number;
  distanceTraveled: number;
  isInPit: boolean;
  currentSectorIndex: number;
  lastSectorTime: number;
  numberOfLaps: number;
  tyreCompound: string;
  replayTimeMultiplier: number;
  normalizedCarPosition: number;
  activeCars: number;
  carCoordinates: Array<number[]>;
  carID: number[];
  playerCarID: number;
  penaltyTime: number;
  flag:
    | 'none'
    | 'blue'
    | 'yellow'
    | 'black'
    | 'white'
    | 'checkered'
    | 'penalty'
    | 'green'
    | 'orange';
  penaltyType: 'none' | 'dt' | 'sg-10' | 'sg-20' | 'sg-30' | 'dsq' | 'rbl';
  penaltyReason:
    | 'none'
    | 'cut'
    | 'pit-speed-limit'
    | 'ignored-mandatory-pit'
    | 'trolling'
    | 'pit-entry'
    | 'pit-exit'
    | 'wrong-way'
    | 'ignored-driver-stint'
    | 'exceeded-driver-stint-limit'
    | 'post-race-time';
  isIdealLineOn: boolean;
  isInPitLane: boolean;
  surfaceGrip: number;
  isMandatoryPitDone: boolean;
  windSpeed: number;
  windDirection: number;
  isSetupMenuVisible: boolean;
  mainDisplayIndex: number;
  secondaryDisplayIndex: number;
  tc: number;
  tcCut: number;
  engineMap: number;
  abs: number;
  fuelXLap: number;
  rainLights: boolean;
  flashingLights: boolean;
  lightsStage: number;
  exhaustTemperature: number;
  wiperLV: number;
  driverStintTotalTimeLeft: number;
  driverStintTimeLeft: number;
  rainTyres: boolean;
  sessionIndex: number;
  usedFuel: number;
  deltaLapTime: string;
  iDeltaLapTime: number;
  estimatedLapTime: string;
  iEstimatedLapTime: number;
  isDeltaPositive: boolean;
  iSplit: number;
  isValidLap: boolean;
  fuelEstimatedLaps: number;
  trackStatus: string;
  missingMandatoryPits: number;
  clock: number;
  directionLightsLeft: boolean;
  directionLightsRight: boolean;
  isGlobalYellow: boolean;
  isGlobalYellow1: boolean;
  isGlobalYellow2: boolean;
  isGlobalYellow3: boolean;
  isGlobalWhite: boolean;
  isGlobalGreen: boolean;
  isGlobalChequered: boolean;
  isGlobalRed: boolean;
  mfdTyreSet: number;
  mfdFuelToAdd: number;
  mfdTyrePressureLF: number;
  mfdTyrePressureRF: number;
  mfdTyrePressureLR: number;
  mfdTyrePressureRR: number;
  trackGripStatus: 'green' | 'fast' | 'optimum' | 'greasy' | 'damp' | 'wet' | 'flooded';
  rainIntensity: RainIntensity;
  rainIntensityIn10min: RainIntensity;
  rainIntensityIn30min: RainIntensity;
  currentTyreSet: number;
  strategyTyreSet: number;
}

export interface Static {
  smVersion: string;
  acVersion: string;
  numberOfSessions: number;
  numCars: number;
  carModel: string;
  track: string;
  playerFirstName: string;
  playerLastName: string;
  playerShortName: string;
  sectorCount: number;
  maxTorque: number;
  maxPower: number;
  maxRPM: number;
  maxFuel: number;
  suspensionMaxTravel: number[];
  tyreRadius: number[];
  maxTurboBoost: number;
  deprecated_1: number;
  deprecated_2: number;
  isPenaltiesEnabled: boolean;
  aidFuelRate: number;
  aidTireRate: number;
  aidMechanicalDamage: number;
  isAllowedTyreBlankets: boolean;
  isAidStabilityEnabled: boolean;
  isAidAutoClutchEnabled: boolean;
  isAidAutoBlipEnabled: boolean;
  hasDRS: boolean;
  hasERS: boolean;
  hasKERS: boolean;
  kersMaxJ: number;
  engineBrakeSettingsCount: number;
  ersPowerControllerCount: number;
  trackSplineLength: number;
  trackConfiguration: string;
  ersMaxJ: number;
  isTimedRace: boolean;
  hasExtraLap: boolean;
  carSkin: string;
  reversedGridPositions: number;
  pitWindowStart: number;
  pitWindowEnd: number;
  isOnline: boolean;
  dryTyresName: string;
  wetTyresName: string;
}